Remove sync headers from public ObjectStore headers
This means that no sync headers need to be included in the exported
packages.

The definitions that are needed from OS headers are collected
in src/realm/sync/config.hpp that should be exported.
