+++ src/realm/sync/changeset_cooker.hpp
@@ -1,5 +1,5 @@
 
-#include <realm/sync/history.hpp>
+#include <realm/sync/config.hpp>
 
 #ifndef REALM_SYNC_CHANGESET_COOKER_HPP
 #define REALM_SYNC_CHANGESET_COOKER_HPP
@@ -8,7 +8,7 @@ namespace realm {
 namespace sync {
 
 /// Copy raw changesets unmodified.
-class TrivialChangesetCooker : public ClientReplication::ChangesetCooker {
+class TrivialChangesetCooker : public ChangesetCooker {
 public:
     bool cook_changeset(const Group&, const char* changeset, std::size_t changeset_size,
                         util::AppendBuffer<char>&) override;
