+++ src/realm/sync/noinst/client_impl_base.cpp
@@ -234,8 +234,8 @@ void Connection::cancel_reconnect_delay()
 Connection::Connection(ClientImplBase& client, std::string logger_prefix, ProtocolEnvelope protocol,
                        std::string address, port_type port, bool verify_servers_ssl_certificate,
                        Optional<std::string> ssl_trust_certificate_path,
-                       std::function<SSLVerifyCallback> ssl_verify_callback, Optional<ProxyConfig> proxy_config,
-                       ReconnectInfo reconnect_info)
+                       std::function<SSLVerifyCallback> ssl_verify_callback,
+                       Optional<SyncConfig::ProxyConfig> proxy_config, ReconnectInfo reconnect_info)
     : logger{std::move(logger_prefix), client.logger} // Throws
     , m_client{client}
     , m_read_ahead_buffer{} // Throws
@@ -721,7 +721,7 @@ void Connection::initiate_resolve()
     const port_type& port = m_proxy_config ? m_proxy_config->port : m_port;
 
     if (m_proxy_config) {
-        logger.detail("Using %1 proxy", m_proxy_config->type); // Throws
+        // logger.detail("Using %1 proxy", m_proxy_config->type); // Throws
     }
 
     logger.detail("Resolving '%1:%2'", address, port); // Throws
