+++ test/object-store/sync/sync_manager.cpp
@@ -18,7 +18,7 @@
 
 #include "sync_test_utils.hpp"
 
-#include <realm/object-store/sync/sync_config.hpp>
+#include <realm/sync/config.hpp>
 #include <realm/object-store/sync/sync_manager.hpp>
 #include <realm/object-store/sync/sync_user.hpp>
 #include "sync/session/session_util.hpp"
